Extra Functionality
Becuase of the Programmable Radio set it was easier to flash existing hardware with a simpler program and support an extra receive data layer than it is to generate an additional set of firmware that modifies the internal AT command post bootloader of the XBEEs. It appears that this message may also contain some interesting information. This should help keep consistency of radios and not have to worry about types received.